move the remaining code to otter.convergence.composition and otter.convergence.test_composition.
